Split animation definitions
